feat: add dependency merging
