+++ packages/saber/lib/cli-commands/dev.js
@@ -1,0 +1,31 @@
+const { setNodeEnv, handleError } = require('./utils')
+
+module.exports = function(cli) {
+  cli
+    .command('[app]', 'Run the application in dev mode', {
+      ignoreOptionDefaultValue: true
+    })
+    .alias('dev')
+    .option('--lazy', 'Enable lazy page compilation')
+    .option('--port <port>', 'Server port', { default: 3000 })
+    .option('--host <host>', 'Server host', { default: '0.0.0.0' })
+    .action((cwd = '.', options) => {
+      setNodeEnv('development')
+
+      const { host, port, lazy } = options
+      delete options.host
+      delete options.port
+      delete options.lazy
+      return require('..')(Object.assign({ cwd, dev: true }, options), {
+        server: {
+          host,
+          port
+        },
+        build: {
+          lazy
+        }
+      })
+        .serve()
+        .catch(handleError)
+    })
+}
