+++ packages/saber/lib/cli-commands/index.js
@@ -1,0 +1,5 @@
+const commands = ['dev', 'build', 'serve', 'eject']
+
+module.exports = function(cli) {
+  commands.forEach(command => require(`./${command}`)(cli))
+}
